fix: move factory prompt run helpers to pkg/config/factoryrun
Keep pkg/config at the 15-file package limit by relocating portable
factory path and prompt-run helpers into a subpackage, and fold
handlingBehavior validation back into config_validator_topology.go.
