+++ pkg/config/factoryrun/factory_config_path.go
@@ -1,4 +1,4 @@
-package config
+package factoryrun
 
 import (
 	"fmt"
